百科用户

+++ src/main/java/com/huotu/huobanplus/sns/model/AppWikiListModel.java
@@ -43,7 +43,7 @@ public class AppWikiListModel {
     /**
      * 用户等级
      */
-    private Integer userLevel;
+    private Long userLevel;
     /**
      * 文章发表时间
      */
